requirements.txt
Django==1.8.4
Pillow==2.9.0
django-crispy-forms==1.5.1
psycopg2==2.6.1
pyserial==2.7
